+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.2.0] - 2025-12-30
+
+### cli
+
+- feat: add support for bidirectional XML ⇄ JSON format conversion.
+
+### webapp
+
+- feat: add XML ⇄ JSON format conversion tool.
+- fix: fix styling issues in webapp for better user experience.
+
 ## [v1.1.2] - 2025-12-26
 
 ### webapp